tasks/osint/time-train/deploy/requirements.txt
httpx ~= 0.23.3
python-telegram-bot==20.1